Reduce sources tcp and csv and sinks to file and print (epic feature #69) (#171)
* #70 : Removed sources expcept TCP,Print,CSV.

* #70 : Formatted sink/source lowering files.

* #70 : Added boost-asio header only lib.

* #70 : TCP source now handles domain names.

* #70 : Fixed blocking TCPSource read with timeout.

* #70 : Implemented TCP/CSV integration tests.

* #70 : Formatting and master rebase.

+++ nes-runtime/tests/UnitTests/Runtime/MemoryLayouts/DynamicMemoryLayoutTest.cpp
@@ -68,7 +68,7 @@ TEST_P(DynamicMemoryLayoutTestParameterized, readWriteColumnartestBufferTest)
 {
     for (int i = 0; i < 10; i++)
     {
-        auto testTuple = std::make_tuple((uint16_t)i, true, i * 2.0);
+        auto testTuple = std::make_tuple(static_cast<uint16_t>(i), true, i * 2.0);
         testBuffer->pushRecordToBuffer(testTuple);
         ASSERT_EQ((testBuffer->readRecordFromBuffer<uint16_t, bool, double>(i)), testTuple);
     }
@@ -78,7 +78,7 @@ TEST_P(DynamicMemoryLayoutTestParameterized, iteratetestBufferTest)
 {
     for (int i = 0; i < 10; i++)
     {
-        testBuffer->pushRecordToBuffer(std::make_tuple(42_u16, true, 42 * 2.0));
+        testBuffer->pushRecordToBuffer(std::make_tuple(static_cast<uint16_t>(42), true, 42 * 2.0));
     }
 
     for (auto tuple : *testBuffer)
